Refactor original parser impl into its own module
